bugfixes in combine_expression
Don't replace '+' with '-' in sub_expression of 'else' branch as it cause wrong result in operation such as "a / (b + c)". The fact it's identical to branch " operator == '-' " lead me to think it's a duplication error.
Additionally, i don't have a django 1.x handy to test but i think "sub_expressions[1] = str.replace('+', '-')" don't work ? It appear to be an artifact of a previous commit when the variable "str" existed and is also redundant with the first branch.